src/requirements.txt
alembic==1.9.2
asyncpg==0.26.0
fastapi==0.89.1
gunicorn==20.1.0
psycopg2-binary==2.9.3
pydantic==1.10.4
python-dotenv==0.21.1
SQLAlchemy==1.4.41
SQLAlchemy-Utils==0.38.3
uvicorn==0.20.0
aiohttp==3.8.3
pytest-asyncio==0.20.2
pytest==7.2.1
pytest-env==0.8.1
aioredis==2.0.1
pbkdf2==1.3
PyJWT==2.6.0
python-multipart==0.0.5
aiofiles==0.6.0